Fixes #2 Support Auto-Associated Accounts
Updated the pre-distribution plan check algorithm to allow
accounts that are not explicitly associated with the token to be
included in the distribution plan.  If the algorithm sees an
account without a pre-existing balance or indication of explicit
association, it will still include the account in the distribution
plan; but will issue a warning.  The user must acknowledge
that transfers to these accounts may fail if they have not
enabled auto-association.  If association fails it will record the
problem in the list of errors and export the status of
TOKEN_NOT_ASSOCIATED_TO_ACCOUNT in the output csv file
(assuming the run is the final one creating enough signatures
to try to make the distribution).

Also added prettier formatting to test code and lengthened
the allowed line size beyond 80 charcters.