Fixes #1 Retry when Hedera Network Unavailable
Added a feature to recognize the condition
when all of the hedera nodes are
temporarily unreachable when trying to
start a distribution.  If this temporary
condition occurs the user is presented with
a “Try Again” button to restart the step
from the beginning.  It will allow the user
to retry as many times as necessary to
establish a connection to a set of nodes.

Also improved the node selection
algorithm to increase the chance of finding
enough adequately responding nodes
when the network is under load.