Fixes #9 and #7  Processing UI Updates
Updated the user interface displaying distribution processing
progress to be easier to understand.  When completed the
user interface distinguishes between the final states of
Completed, Scheduled and Failed.  It now hides the detailed
list of in-progress distributions by default, but can still be
displayed by clicking on a view details link.  The updated
status values also extend to the CSV output file as well.