Change to DragonGlass & HashScan Explorers
Fixes #39  Updated user interface to swap
Kabuto explorer links to DragonGlass &
HashScan explorer links.

This introduces an additional dependence
on a mirror node for the target network.
The algorithms first query the mirror node
for basic information to display to the user
which additionally help formulate links to
the various explorers.

Not all explorers support all types of entitites
and transactions created by this application.