Fix for Network 0.55 Account Balance Throttle
The implementation of the balance query throttle
implemented in version 0.55 renders the previous
version of this application unusable.  Checks for
valid token reciever accounts have been moved from
querying network nodes to the mirror node.  This
will slow the validation process, however even the
smallest of distributions have shown to be negativly
impacted by this change in network behavior.